sim: change the type of xcpt_reg_t

+++ arch/sim/include/setjmp.h
@@ -27,6 +27,7 @@
 
 #include <nuttx/config.h>
 #include <nuttx/compiler.h>
+#include <sys/types.h>
 
 /****************************************************************************
  * Pre-processor Definitions
@@ -156,7 +157,7 @@
 
 #ifndef __ASSEMBLY__
 
-typedef unsigned long xcpt_reg_t;
+typedef size_t xcpt_reg_t;
 typedef xcpt_reg_t jmp_buf[XCPTCONTEXT_REGS];
 
 #endif
